+++ point.h
@@ -1,11 +1,12 @@
 typedef struct
 {
   int id;
-  float x1, x2, a, b, x, y;
-  float* distances;
+  double x1, x2, a, b, x, y;
+  double* distances;
 } Point;
 
 Point *allocatePoint(int distancesSize);
 void deallocatePoint(Point* p);
 void setPosition(Point* p, float t);
 float calculateDistanceBetweenPoints(Point* p1, Point* p2);
+
